% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.scaling.r
\name{am.range01}
\alias{am.range01}
\title{Scale to 0-1 range}
\usage{
am.range01(x)
}
\arguments{
\item{x}{numeric}
}
\value{
Numeric
}
\description{
Scale numeric vector to 0-1 range. Source: http://stackoverflow.com/questions/5468280/scale-a-series-between-two-points
}
\seealso{
`%*na%`
}

