% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.tdiff.r
\name{am.tdiff}
\alias{am.tdiff}
\title{Scale to 0-1 range}
\usage{
am.tdiff(times, num_per_segment = 60)
}
\arguments{
\item{times}{let's try what is it}

\item{num_per_segment}{another argument}
}
\value{
Time (?)
}
\description{
Scale numeric vector to 0-1 range. Source: http://stackoverflow.com/questions/5468280/scale-a-series-between-two-points
}
\seealso{
...
}

