% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.operator.r
\name{\%fin\%}
\alias{\%fin\%}
\title{Asymmetric matching two vectors}
\usage{
x \%fin\% srch
}
\arguments{
\item{x}{vector}

\item{srch}{vector}
}
\value{
Logical vector of the length of x
}
\description{
Asymmetric matching two vectors
}
\examples{
`\%fin\%`(c("a"," b","c","d","e"),c("a","d","e"))
}
\seealso{
`%+na%`, `%*na%`
}

