% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.operator.r
\name{\%*na\%}
\alias{\%*na\%}
\title{Multiple excluding NAs}
\usage{
x \%*na\% y
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
Numeric
}
\description{
Multiple two numbers excluding NAs. Not vectorized
}
\seealso{
`%+na%`
}

