% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.tdiff.r
\name{am.tdiff}
\alias{am.tdiff}
\title{Time difference}
\usage{
am.tdiff(times, num_per_segment = 60)
}
\arguments{
\item{times}{let's try what is it}

\item{num_per_segment}{another argument}
}
\value{
Time (?)
}
\description{
It does something
Source: Kaggle (?)
Author: potterzot https://www.kaggle.com/potterzot (?)
Reference: https://www.kaggle.com/potterzot/how-much-did-it-rain-ii/features/run/83496
}
\seealso{
...
}

