% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.smath.r
\name{amm.is.wholenumber}
\alias{amm.is.wholenumber}
\title{Test if is whole number}
\usage{
amm.is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric}

\item{tol}{numeric: tolerance, treshold}
}
\value{
Logical
}
\description{
Test if the number is the whole number
}
\seealso{
...
}

