% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.outersect.r
\name{amm.outersect}
\alias{amm.outersect}
\title{Outersect (vector XOR)}
\usage{
amm.outersect(x, y, ...)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}

\item{...}{other vectors (optional)}
}
\value{
Vector containing only unique values (those do not have duplicates in other input vectors)
}
\description{
Outersect (vector XOR).
Author: @tonybreyal
Reference: https://tonybreyal.wordpress.com/2011/11/29/outersect-the-opposite-of-rs-intersect-function/
}
\seealso{
...
}

