% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.sort.named.r
\name{amm.sort.named}
\alias{amm.sort.named}
\title{Test if is whole number}
\usage{
amm.sort.named(vc.v, vc.n, qnt = NA)
}
\arguments{
\item{vc.v}{vector to be named}

\item{qnt}{numeric length of vector to be returned (You may cut vector to less length if you prefer)}

\item{tol}{vector of names}
}
\value{
Logical
}
\description{
Test if the number is the whole number
}
\seealso{
...
}

